+++ Taskify.API.Test/TasksUnitTests/DeleteTasksUnitTest.cs
@@ -7,34 +7,34 @@ using System.Text;
 using System.Threading.Tasks;
 using Taskify.API.Controllers;
 
-namespace Taskify.API.Test.UnitTests.Tasks.Get
+namespace Taskify.API.Test.UnitTests
 {
-    public class GetTaskByDateUnitTest : IClassFixture<TasksUnitTestController>
+    public class DeleteTasksUnitTest : IClassFixture<TasksUnitTestController>
     {
         private readonly TaskController _controller;
 
-        public GetTaskByDateUnitTest(TasksUnitTestController controller)
+        public DeleteTasksUnitTest(TasksUnitTestController controller)
         {
             _controller = new TaskController(controller.repository);
         }
 
         [Fact]
-        public async Task GetTaskByDate_ReturnsAllItems()
+        public async Task DeleteTask_ReturnsOk()
         {
             // act
-            var result = await _controller.FilterTasksByDate(DateTime.Now);
+            var result = await _controller.RemoveTask(1);
 
             // assert
             result.Result.Should().BeOfType<OkObjectResult>()
-                .Which.Value.Should().BeAssignableTo<IEnumerable<Taskify.API.Models.Tasks>>()
+                .Which.Value.Should().BeAssignableTo<Models.Tasks>()
                 .And.NotBeNull();
         }
 
         [Fact]
-        public async Task GetTaskByDate_ReturnsNotFound()
+        public async Task DeleteTask_ReturnsNotFound()
         {
             // act
-            var result = await _controller.FilterTasksByDate(DateTime.MinValue);
+            var result = await _controller.RemoveTask(0);
 
             // assert
             result.Result.Should().BeOfType<NotFoundResult>()
