+++ Taskify.API.Test/TasksUnitTests/Get/GetTaskByDateUnitTest.cs
@@ -7,34 +7,34 @@ using System.Text;
 using System.Threading.Tasks;
 using Taskify.API.Controllers;
 
-namespace Taskify.API.Test.UnitTests.Tasks.Get
+namespace Taskify.API.Test.UnitTests.Get
 {
-    public class GetTaskByIdUnitTest : IClassFixture<TasksUnitTestController>
+    public class GetTaskByDateUnitTest : IClassFixture<TasksUnitTestController>
     {
         private readonly TaskController _controller;
 
-        public GetTaskByIdUnitTest(TasksUnitTestController controller)
+        public GetTaskByDateUnitTest(TasksUnitTestController controller)
         {
             _controller = new TaskController(controller.repository);
         }
 
         [Fact]
-        public async Task GetTaskById_WhenCalled_ReturnsItem()
+        public async Task GetTaskByDate_ReturnsAllItems()
         {
             // act
-            var result = await _controller.GetTaskById(1);
+            var result = await _controller.FilterTasksByDate(DateTime.Now);
 
             // assert
             result.Result.Should().BeOfType<OkObjectResult>()
-                .Which.Value.Should().BeAssignableTo<Taskify.API.Models.Tasks>()
+                .Which.Value.Should().BeAssignableTo<IEnumerable<Models.Tasks>>()
                 .And.NotBeNull();
         }
 
         [Fact]
-        public async Task GetTaskById_WhenCalled_ReturnsNotFound()
+        public async Task GetTaskByDate_ReturnsNotFound()
         {
             // act
-            var result = await _controller.GetTaskById(999999);
+            var result = await _controller.FilterTasksByDate(DateTime.MinValue);
 
             // assert
             result.Result.Should().BeOfType<NotFoundResult>()
