+++ Taskify.API.Test/TasksUnitTests/Get/GetTaskByStatusUnitTest.cs
@@ -7,7 +7,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Taskify.API.Controllers;
 
-namespace Taskify.API.Test.UnitTests.Tasks.Get
+namespace Taskify.API.Test.UnitTests.Get
 {
     public class GetTaskByStatusUnitTest : IClassFixture<TasksUnitTestController>
     {
@@ -26,19 +26,19 @@ namespace Taskify.API.Test.UnitTests.Tasks.Get
 
             // assert
             result.Result.Should().BeOfType<OkObjectResult>()
-                .Which.Value.Should().BeAssignableTo<IEnumerable<Taskify.API.Models.Tasks>>()
+                .Which.Value.Should().BeAssignableTo<IEnumerable<Models.Tasks>>()
                 .And.NotBeNull();
         }
 
         [Fact]
-        public async Task  GetTaskByStatus_ReturnsAllItems_NotDone()
+        public async Task GetTaskByStatus_ReturnsAllItems_NotDone()
         {
             // act
             var result = await _controller.FilterTasksByStatus(false);
 
             // assert
             result.Result.Should().BeOfType<OkObjectResult>()
-                .Which.Value.Should().BeAssignableTo<IEnumerable<Taskify.API.Models.Tasks>>()
+                .Which.Value.Should().BeAssignableTo<IEnumerable<Models.Tasks>>()
                 .And.NotBeNull();
         }
     }
