+++ Taskify.API.Test/TasksUnitTests/PutTasksUnitTest.cs
@@ -7,7 +7,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Taskify.API.Controllers;
 
-namespace Taskify.API.Test.UnitTests.Tasks
+namespace Taskify.API.Test.UnitTests
 {
     public class PutTasksUnitTest : IClassFixture<TasksUnitTestController>
     {
@@ -22,12 +22,12 @@ namespace Taskify.API.Test.UnitTests.Tasks
         public async Task PutTask_ReturnsOk()
         {
             // act
-            var result = await _controller.UpdateTask(new Taskify.API.Models.Tasks 
+            var result = await _controller.UpdateTask(new Models.Tasks
             { Id = 1, Titulo = "Task 1", Descricao = "Task 1", DataVencimento = DateTime.Now, Concluida = false });
 
             // assert
             result.Result.Should().BeOfType<OkObjectResult>()
-                .Which.Value.Should().BeAssignableTo<Taskify.API.Models.Tasks>()
+                .Which.Value.Should().BeAssignableTo<Models.Tasks>()
                 .And.NotBeNull();
         }
     }
