--- Taskify.API.Test/UnitTests/Auth/RegisterUnitTest.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Taskify.API.Test.UnitTests.Auth
-{
-    internal class RegisterUnitTest
-    {
-    }
-}
